requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.5
pypdf==5.1.0
langchain==0.3.15
langchain-openai==0.3.2
langchain-community==0.3.15
faiss-cpu==1.9.0.post1